fix : addEntity #51
