--- backend/src/main/java/com/kh/ct/Attendance/entity/Attendance.java
@@ -1,41 +1,0 @@
-package com.kh.ct.Attendance.entity;
-
-import com.kh.ct.Common.entity.BaseTimeEntity;
-import com.kh.ct.Common.entity.CommonEnums;
-import com.kh.ct.Member.entity.Emp;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Attendance extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long attendanceId;
-
-    @Column(nullable = false)
-    private LocalDate attendanceDate;
-
-    private LocalTime inTime;
-
-    private LocalTime outTime;
-
-
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private CommonEnums.AttendanceStatus attendanceStatus;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "emp_id")
-    private Emp empId;
-
-}
