--- backend/src/main/java/com/kh/ct/Common/entity/Code.java
@@ -1,26 +1,0 @@
-package com.kh.ct.Common.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Code extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long codeId;
-
-    @Column(nullable = false, length = 100)
-    private String codeName;
-
-    @OneToMany(mappedBy = "codeId", fetch = FetchType.LAZY,
-            cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<CodeDetail> codeDetails = new ArrayList<>();
-}
