--- backend/src/main/java/com/kh/ct/Common/entity/File.java
@@ -1,27 +1,0 @@
-package com.kh.ct.Common.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class File extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long fileId;
-
-    @Column(nullable = false)
-    private String fileName;
-
-    @Column(nullable = false)
-    private String fileOriName;
-
-    @Column(nullable = false)
-    private String path;
-
-    private Long size;
-}
