+++ backend/src/main/java/com/kh/ct/Domain/Member/entity/PasswordCode.java
@@ -1,0 +1,35 @@
+package com.kh.ct.Domain.Member.entity;
+
+import com.kh.ct.Global.common.CommonEnums;
+import com.kh.ct.Global.entity.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class PasswordCode extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long PasswordCodeId;
+
+    @JoinColumn(nullable = false, name = "emp_id")
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    private Emp empId;
+
+    @Column(length = 200)
+    private String passwordCode;
+
+    private LocalDateTime codeExpiresDate;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private CommonEnums.CommonStatus leaveApplyStatus;
+
+
+}
