+++ backend/src/main/java/com/kh/ct/Domain/Member/entity/PasswordToken.java
@@ -1,0 +1,36 @@
+package com.kh.ct.Domain.Member.entity;
+
+
+import com.kh.ct.Global.common.CommonEnums;
+import com.kh.ct.Global.entity.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class PasswordToken extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long PasswordTokenId;
+
+    @JoinColumn(nullable = false, name = "emp_id")
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    private Emp empId;
+
+    @Column(length = 200)
+    private String passwordToken;
+
+    private LocalDateTime tokenExpiresDate;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private CommonEnums.CommonStatus leaveApplyStatus;
+
+
+}
