+++ backend/src/main/java/com/kh/ct/Domain/Schedule/entity/Airport.java
@@ -1,0 +1,35 @@
+package com.kh.ct.Domain.Schedule.entity;
+
+import com.kh.ct.Domain.Member.entity.AirlineAirport;
+import com.kh.ct.Global.entity.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Airport extends BaseTimeEntity{
+    @Id
+    private String airportCode;
+
+    @Column(length = 100)
+    private String airportName;
+
+    @Column(length = 100)
+    private String countryName;
+
+    @Column(length = 100)
+    private String timezone;
+
+    @Column(length = 100)
+    private String cityName;
+
+    @OneToMany(mappedBy = "airportId", fetch = FetchType.LAZY,
+            cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<AirlineAirport> airlineAirports = new ArrayList<>();
+}
