+++ backend/src/main/java/com/kh/ct/Global/entity/File.java
@@ -1,0 +1,27 @@
+package com.kh.ct.Global.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class File extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long fileId;
+
+    @Column(nullable = false)
+    private String fileName;
+
+    @Column(nullable = false)
+    private String fileOriName;
+
+    @Column(nullable = false)
+    private String path;
+
+    private Long size;
+}
