--- backend/src/main/java/com/kh/ct/Member/entity/Airline.java
@@ -1,37 +1,0 @@
-package com.kh.ct.Member.entity;
-
-import com.kh.ct.Common.entity.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Airline extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long airlineId;
-
-    @Column(length = 100)
-    private String airlineName;
-
-    @Column(length = 100)
-    private String theme;
-
-    @Column(length = 50)
-    private String mainNumber;
-
-    @Column(length = 255)
-    private String airlineAddress;
-
-    @Column(length = 500)
-    private String airlineDesc;
-
-    @JoinColumn(name = "airline_apply_id")
-    @OneToOne(fetch = FetchType.LAZY, optional = false)
-    private AirlineApply airlineApplyId;
-}
