--- backend/src/main/java/com/kh/ct/Member/entity/EmpHistory.java
@@ -1,30 +1,0 @@
-package com.kh.ct.Member.entity;
-
-import com.kh.ct.Common.entity.BaseTimeEntity;
-import com.kh.ct.Common.entity.File;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class EmpHistory extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long empHistoryId;
-
-    @JoinColumn(name = "emp_id")
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    private Emp empId;
-
-    @Column(length = 50)
-    private String job;
-
-    @Column(length = 50)
-    private String historyDepartment;
-}
