--- backend/src/main/java/com/kh/ct/Schedule/entity/AllSchedule.java
@@ -1,20 +1,0 @@
-package com.kh.ct.Schedule.entity;
-
-import com.kh.ct.Common.entity.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AllSchedule extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long scheduleId;
-
-    @Column(length = 20)
-    private String scheduleCode; // FLIGHT, COUNSEL, EXERCISE, REST 같은 형태 추천
-}
